% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies_functions.R
\name{appendvar}
\alias{appendvar}
\title{Append variables function
appends variable in df1 with variable from df2 by aligning common group column that exists in both dfs}
\usage{
appendvar(var1, var2, var3)
}
\arguments{
\item{var1}{variable from df1 to align with df2}

\item{var2}{variable from df2 to align with df1}

\item{var3}{value column from df2 to append to df1}
}
\value{
dataframe with appended variable
}
\description{
Append variables function
appends variable in df1 with variable from df2 by aligning common group column that exists in both dfs
}
\examples{
lat = appendvar(df$site, sites$site, sites$latitude)
}
