% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies_functions.R
\name{convert_classes_daily}
\alias{convert_classes_daily}
\title{Convert column classes of daily data
This function takes daily data and converts columns to appropriate class (numeric or character)}
\usage{
convert_classes_daily(df)
}
\arguments{
\item{df}{Dataframe}
}
\value{
Dataframe with appropriate column classes
}
\description{
Convert column classes of daily data
This function takes daily data and converts columns to appropriate class (numeric or character)
}
\note{
This function is meant to be applied to flagged and/or data
}
\examples{
df <- convert_classes_daily(df)
}
