% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies_functions.R
\name{flag_data}
\alias{flag_data}
\title{Data flagging function}
\usage{
flag_data(df)
}
\arguments{
\item{df}{raw climate dataframe to be flagged}
}
\value{
flagged cliamte dataframe
}
\description{
Data flagging function
}
\examples{
df <- flag_data(df)

}
