% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies_functions.R
\name{import_site_data}
\alias{import_site_data}
\title{Import data for select sites}
\usage{
import_site_data(site, data_path = "./")
}
\arguments{
\item{site}{A character vector of site names}

\item{data_path}{A character string specifying path to directory}
}
\value{
A data frame containing the imported data
}
\description{
Import data for select sites
}
\examples{
# Set data_path directory
data_path <- "C:/Users/path/to/your/data"
# Define sites
site <- c("Yellowknife","Scotty Creek","Deline","Blueberry")
# Import data
data <- import_site_data(site, data_path)``
}
