% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_station_data.R
\name{map_station_data}
\alias{map_station_data}
\title{Map station data}
\usage{
map_station_data(
  parameter = "snow",
  select_year = 2023,
  months = c(1:12),
  start_year = NULL,
  max_missing_days = 20,
  percent_of_normal = TRUE,
  years_of_record = 5,
  circle_radius = 10,
  water_year = TRUE,
  filename = NA
)
}
\arguments{
\item{parameter}{The parameter to map}

\item{select_year}{The year to map}

\item{months}{The months to map}

\item{start_year}{The first year to map}

\item{max_missing_days}{The maximum number of missing days over the time period}

\item{percent_of_normal}{Logical. Determines whether map produces percent of normal or absolute value}

\item{years_of_record}{Numeric. The minimum number of years of record for station to appear}

\item{circle_radius}{Numeric. The radius of the circle in leaflet}

\item{water_year}{Logical. Determines whether map uses a water year or calendar year}

\item{filename}{Character. The name of the file to save}
}
\value{
A map of all stations in the NWT
}
\description{
Produce a map of all station data
}
