% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies_functions.R
\name{parameter_name}
\alias{parameter_name}
\title{Function to determine parameter name based on different inputs}
\usage{
parameter_name(parameter)
}
\arguments{
\item{parameter}{the input parameter}
}
\value{
standardized parameter name
}
\description{
Function to determine parameter name based on different inputs
}
