% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies_functions.R,
%   R/replace_coordinates.R
\name{replace_coordinates}
\alias{replace_coordinates}
\title{Amend coordinates of select site}
\usage{
replace_coordinates(df, site, lat, lon)

replace_coordinates(df, site, lat, lon)
}
\arguments{
\item{df}{dataframe}

\item{site}{Site of coordinates to change (in quotations)}

\item{lat}{New latitude value of the site}

\item{lon}{New longitude value of the site}
}
\value{
dataframe with updated coordinates for input site

dataframe with updated coordinates for input site
}
\description{
Amend coordinates of select site

Amend coordinates of select site
}
\examples{
df <- replace_coordinates(df, "Fort McPherson", 67.45, -134.88)
# Example usage:
df <- replace_coordinates(df, "Fort McPherson", 67.45, -134.88)
}
