% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies_functions.R
\name{summarize_data}
\alias{summarize_data}
\title{Summarize data into daily values
This function takes hourly or half-hourly data and summarizes it into daily data}
\usage{
summarize_data(df)
}
\arguments{
\item{df}{Dataframe to be summarized}
}
\value{
Dataframe of daily data
}
\description{
Summarize data into daily values
This function takes hourly or half-hourly data and summarizes it into daily data
}
\note{
This function relies on existence of following columns: station_name, merged_name, lat, lon, elev, year, JD, month, day, date
}
\examples{
df <- summarize_data(df)
}
